+++ parallel_prog/lib/pagerank.h
@@ -7,17 +7,19 @@
 #include "proto/page.pb.h"
 
 std::vector<int> ExploreDanglingPages(const std::vector<int> &out_link_cnts);
-std::vector<double> InitPr(int page_cnt);
+std::vector<long double> InitPr(int page_cnt);
 void AddPagesPr(
 		const std::vector<pr::Page> &pages,
 		const std::vector<int> &out_link_cnts,
-		const std::vector<double> &old_pr,
-		std::vector<double> &new_pr);
+		const std::vector<long double> &old_pr,
+		std::vector<long double> &new_pr);
 void AddDanglingPagesPr(
-		double damping_factor,
-		const std::vector<int> &dangling_pages, 
-		const std::vector<double> &old_pr,
-		std::vector<double> &new_pr);
-void AddRandomJumpsPr(double damping_factor, std::vector<double> &new_pr);
+		long double damping_factor,
+		const std::vector<int> &dangling_pages,
+		const std::vector<long double> &old_pr,
+		std::vector<long double> &new_pr);
+void AddRandomJumpsPr(
+		long double damping_factor,
+		std::vector<long double> &new_pr);
 
 #endif  // PAGERANK_H_
